created AddressRepo.

+++ src/main/java/com/example/hibarnet_testing/repositories/UserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface UserRepo extends CrudRepository<User,Long> {
+public interface UserRepo extends CrudRepository<User,Integer> {
     List<User> findByName(String name);
 }
